Set up 'copy link' functionality
This might be the last major commit to this application. I understand that it took longer than expected to finish building it out-partially due to exams disrupting and all. but regardless. Lots of things I did here were new to me and I really enjoyed creating this simple fullstack web app